# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 1.0.0
  description: |
    Tree Friend E-Commerce API.

    ## Versioning

    The API is mounted under two prefixes:
      - `/api/v1` — the canonical, versioned path. New clients should use this.
      - `/api`     — backward-compat alias for existing callers. Same handlers,
                     transparently aliased.

    When a v2 is needed, a new router will be mounted under `/api/v2` and
    `/api/v1` will be left untouched. The unversioned `/api` alias will be
    deprecated (Deprecation header + warning log) once all callers have
    migrated to `/api/v1`.
servers:
  - url: /api/v1
    description: Versioned API path (canonical)
  - url: /api
    description: Unversioned API path (backward-compat alias)
...
paths:
...
  /cart/items:
    post:
      operationId: addToCart
      tags: [cart]
      summary: Add item to cart
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddToCartBody"
      responses:
        "200":
          description: Updated cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
  /cart/items/{id}:
    put:
      operationId: updateCartItem
      tags: [cart]
      summary: Update cart item quantity
      description: >-
        id is the cart_items row's own primary key, not a variantId --
        a seller-listing cart line has no variantId, so the row id is the
        only identifier that addresses both line kinds (plan doc §2, §7;
        see schema/cart.ts doc comment).
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCartItemBody"
      responses:
        "200":
          description: Updated cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
    delete:
      operationId: removeFromCart
      tags: [cart]
      summary: Remove item from cart
      description: id is the cart_items row's own primary key, not a variantId.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
...
components:
  schemas:
...
    ProductVariant:
      type: object
      properties:
        id:
          type: integer
        productId:
          type: integer
        name:
          type: string
        variantType:
          type: string
        form:
          type: ["string", "null"]
        price:
          type: number
        discountPrice:
          type: ["number", "null"]
        stock:
          type: integer
        deliveryCharge:
          type: number
        sku:
          type: ["string", "null"]
      required:
        - id
        - productId
        - name
        - variantType
        - price
        - stock
        - deliveryCharge
...
    CartItemProduct:
      type: object
      description: Trimmed product info attached to a cart line (not the full Product shape).
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        categoryId:
          type: integer
        description:
          type: string
        images:
          type: array
          items:
            type: string
        averageRating:
          type: number
        reviewCount:
          type: integer
        isFeatured:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - name
        - slug
        - categoryId
        - description
        - images
        - averageRating
        - reviewCount
        - createdAt
    CartItemSeller:
      type: object
      description: Present only on seller_listing-kind cart lines.
      properties:
        id:
          type: integer
        businessName:
          type: string
        nurseryName:
          type: string
        location:
          type: string
        hasVerifiedPaymentConfig:
          type: boolean
          description: >-
            True only if this seller has a seller_payment_configs row with
            isVerified = true. Reflects the seller's live config state, which
            can drift from the listing's own paymentMethod field (e.g. after
            an admin unverifies a seller without touching their listings).
            Checkout uses this, not listing.paymentMethod alone, to decide
            whether bKash can actually be offered for this line.
      required:
        - id
        - businessName
        - nurseryName
        - location
        - hasVerifiedPaymentConfig
    CartItemListing:
      type: object
      description: Present only on seller_listing-kind cart lines.
      properties:
        id:
          type: integer
        form:
          type: ["string", "null"]
        rootType:
          type: ["string", "null"]
        potSize:
          type: ["string", "null"]
        age:
          type: ["string", "null"]
        height:
          type: ["string", "null"]
        condition:
          type: ["string", "null"]
        price:
          type: number
        discountPrice:
          type: ["number", "null"]
        stock:
          type: integer
        availableQuantity:
          type: integer
        deliveryTimeDays:
          type: ["integer", "null"]
        deliveryCharge:
          type: number
        paymentMethod:
          type: string
      required:
        - id
        - price
        - stock
        - availableQuantity
        - deliveryCharge
        - paymentMethod
    CartItem:
      type: object
      description: >-
        A cart line is EITHER kind "variant" (admin-direct, variant +
        variantId populated) OR kind "seller_listing" (marketplace,
        sellerListingId/sellerId/seller/listing populated) -- never both.
        See schema/cart.ts doc comment for the full rationale.
      properties:
        id:
          type: integer
        kind:
          type: string
          enum: [variant, seller_listing]
        productId:
          type: integer
        variantId:
          type: ["integer", "null"]
        sellerListingId:
          type: ["integer", "null"]
        sellerId:
          type: ["integer", "null"]
        quantity:
          type: integer
        variant:
          oneOf:
            - $ref: "#/components/schemas/ProductVariant"
            - type: "null"
        seller:
          oneOf:
            - $ref: "#/components/schemas/CartItemSeller"
            - type: "null"
        listing:
          $ref: "#/components/schemas/CartItemListing"
        product:
          $ref: "#/components/schemas/CartItemProduct"
      required:
        - id
        - kind
        - productId
        - variantId
        - sellerListingId
        - sellerId
        - quantity
        - product
    Cart:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/CartItem"
        subtotal:
          type: number
        discount:
          type: number
        deliveryTotal:
          type: number
        total:
          type: number
      required:
        - items
        - subtotal
        - discount
        - deliveryTotal
        - total
    AddToCartBody:
      type: object
      description: >-
        Specify exactly one of variantId (admin-direct product variant) or
        sellerListingVariantId (marketplace seller listing variant), never
        both, never neither -- confirmed against cart.ts's POST /cart/items
        handler, which rejects the request with 400 if hasVariant ===
        hasListingVariant. sellerListingId is still accepted in the request
        body for backward compat with older clients, but is IGNORED for
        line-creation purposes -- the route derives sellerListingId itself
        from sellerListingVariantId's own FK rather than trusting a
        client-sent value.
      properties:
        productId:
          type: integer
        variantId:
          type: integer
          nullable: true
        sellerListingId:
          type: integer
          nullable: true
          description: Deprecated/ignored -- see description above. Prefer sellerListingVariantId.
        sellerListingVariantId:
          type: integer
          nullable: true
        quantity:
          type: integer
      required:
        - productId
        - quantity
    UpdateCartItemBody:
      type: object
      properties:
        quantity:
          type: integer
      required:
        - quantity
...